index: extract AncestorsBitSet to new module
I'll add more primitive BitSet type there.
